requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cmudict==1.0.32
contourpy==1.3.0
cycler==0.12.1
emoji==2.14.1
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
group-lasso==1.5.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
stanza==1.10.1
sympy==1.13.1
textstat==0.7.7
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
zipp==3.22.0
